+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 10.6
 - Truncate Assignment title to 36 chars in StudentGrades.php
 - CSS Do NOT use global word-break. Use it only on specific elements in stylesheet.css & zresponsive.css
 - CSS responsive Fix list overflow-x scroll inside div.st in stylesheet.css
+- CSS set max-width for City & State select in Address.inc.php
 
 Changes in 10.5.2
 -----------------